+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## dbt 0.18.0 (Release TBD)
 
+## dbt 0.18.0b2 (July 30, 2020)
+
+
 ### Breaking changes
 - Previously, dbt put macros from all installed plugins into the namespace. This version of dbt will not include adapter plugin macros unless they are from the currently-in-use adapter or one of its dependencies [#2590](https://github.com/fishtown-analytics/dbt/pull/2590)